Update dependency injection
