+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Configuration/ChargeLinkBundleSenderConfiguration.cs
@@ -14,10 +14,8 @@
 
 using GreenEnergyHub.Charges.Application.ChargeLinks.MessageHub;
 using GreenEnergyHub.Charges.Application.ChargeLinks.MessageHub.Infrastructure;
-using GreenEnergyHub.Charges.Domain.MarketParticipants;
 using GreenEnergyHub.Charges.Infrastructure.ChargeLinkBundle.Cim;
 using GreenEnergyHub.Charges.Infrastructure.ChargeLinkBundle.MessageHub;
-using GreenEnergyHub.Charges.Infrastructure.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GreenEnergyHub.Charges.FunctionHost.Configuration
@@ -29,12 +27,6 @@ namespace GreenEnergyHub.Charges.FunctionHost.Configuration
             serviceCollection.AddScoped<IChargeLinkBundleSender, ChargeLinkBundleSender>();
             serviceCollection.AddScoped<IChargeLinkBundleCreator, ChargeLinkBundleCreator>();
             serviceCollection.AddScoped<IChargeLinkBundleReplier, ChargeLinkBundleReplier>();
-            serviceCollection.AddScoped<IHubSenderConfiguration>(_ =>
-            {
-                var senderId = EnvironmentHelper.GetEnv("HUB_SENDER_ID");
-                var roleIntText = EnvironmentHelper.GetEnv("HUB_SENDER_ROLE_INT_ENUM_VALUE");
-                return new HubSenderConfiguration(senderId, (MarketParticipantRole)int.Parse(roleIntText));
-            });
             serviceCollection.AddScoped<IChargeLinkCimSerializer, ChargeLinkCimSerializer>();
         }
     }
